+++ queosk/src/main/java/com/bttf/queosk/entity/Restaurant.java
@@ -5,7 +5,7 @@ import com.bttf.queosk.dto.restaurantdto.UpdateRestorantInfoForm;
 import com.bttf.queosk.entity.baseentity.BaseTimeEntity;
 import com.bttf.queosk.enumerate.OperationStatus;
 import com.bttf.queosk.enumerate.RestaurantCategory;
-import com.bttf.queosk.model.userModel.UserRole;
+import com.bttf.queosk.model.usermodel.UserRole;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
