+++ queosk/src/main/java/com/bttf/queosk/model/usermodel/UserStatus.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.model.userModel;
+package com.bttf.queosk.model.usermodel;
 
 public enum UserStatus {
     VERIFIED,
