+++ queosk/src/test/java/com/bttf/queosk/repository/QueueRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.bttf.queosk.repository;
 
-import com.bttf.queosk.model.userModel.UserStatus;
+import com.bttf.queosk.model.usermodel.UserStatus;
 import com.bttf.queosk.entity.Queue;
 import com.bttf.queosk.entity.Restaurant;
 import com.bttf.queosk.entity.User;
