+++ queosk/src/test/java/com/bttf/queosk/repository/UserRepositoryTest.java
@@ -2,7 +2,7 @@ package com.bttf.queosk.repository;
 
 import com.bttf.queosk.entity.baseentity.JpaAuditingConfiguration;
 import com.bttf.queosk.entity.User;
-import com.bttf.queosk.model.userModel.UserStatus;
+import com.bttf.queosk.model.usermodel.UserStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
