+++ queosk/src/test/java/com/bttf/queosk/service/userservice/UserServiceTest.java
@@ -19,7 +19,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 import java.util.Optional;
 
-import static com.bttf.queosk.model.userModel.UserStatus.*;
+import static com.bttf.queosk.model.usermodel.UserStatus.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.hibernate.validator.internal.util.Contracts.assertNotNull;
